Create UseCourseData hook
