+++ src/app/private/course/[courseId]/board/page.tsx
@@ -4,10 +4,9 @@ import Board from "@components/board";
 import Header from "@components/Header";
 import Spinner from "@components/Spinner";
 import { useOfficeHour } from "@hooks/oh/useOfficeHour";
+import { useCourseData } from "@hooks/useCourseData";
 import { Box, Button, Typography } from "@mui/material";
-import { getCourse } from "@services/client/course";
-import { getUsers } from "@services/client/user";
-import { getActiveQuestions, getUserSessionOrRedirect } from "@utils/index";
+import { getActiveQuestions } from "@utils/index";
 import Link from "next/link";
 
 import React from "react";
@@ -20,28 +19,10 @@ interface PageProps {
 
 const Page = (props: PageProps) => {
   const { courseId } = props.params;
-  const user = getUserSessionOrRedirect();
   const { questions } = useOfficeHour();
   const activeQuestions = getActiveQuestions(questions);
 
-  const [loading, setLoading] = React.useState(true);
-  const [isUserTA, setIsUserTA] = React.useState(false);
-
-  React.useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const courseData = await getCourse(courseId);
-        const tasData = await getUsers(courseData.tas);
-        setIsUserTA(tasData.some((ta) => ta.id === user.id));
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, [courseId]);
+  const { loading, isUserTA } = useCourseData({ fetchUsers: false });
 
   if (loading) {
     return (
